requirements.txt
Django==5.2.10
django-environ==0.12.0
django-modeltranslation==0.19.19
django-unfold==0.76.0
Pillow==12.1.0
asgiref==3.11.0
sqlparse==0.5.5
typing_extensions==4.15.0
gunicorn==23.0.0
whitenoise==6.9.0
